Messagesview and propertykeyselectiontree does not handle its default state well (#21)
* Added multi-selection support to ResourceBundleSelectionDialog.

* MessageView and PropertyKeySelectionTree cause NPEs

Initializes more of the view even if the project and resource bundle is
null. Also, a lot of NPE guards

* MessageView and PropertyKeySelectionTree cause NPEs

More NPE guards and a pretty large change to the ResKeyTreeLabelProvider
where code was moved from an explicitly called updateTreeViewer method
to the implicitly called update method.

* MessageView and PropertyKeySelectionTree cause NPEs

Disabled the search bar and the tree if no ResourceBundle is selected.
Also fixed some invalid states. The Memento and ViewState still do not
work properly though. That will be fixed in another issue since I
suspect that there are deeper problems there.